IGI/LR5/requirements.txt
Django>=5.2.1
django-crispy-forms>=2.1
crispy-bootstrap4>=2024.1
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=4.1.0
pytz>=2024.1
Pillow>=11.0
psycopg2-binary>=2.9.9  # Optional, for PostgreSQL support
matplotlib>=3.8.0 